fix: routes with trips can be updated and saved
#67

+++ app_vukwm_bag_delivery/dispatch_routes/route_sheet_cell_gen.py
@@ -232,7 +232,7 @@ def generate_sheet_cells():
         **{
             "Vehicle Id": assigned_jobs["Vehicle Id"]
             + " "
-            + assigned_jobs["Trip Id"].astype(str).str.zfill(2)
+            + assigned_jobs["Trip Id"].astype(int).astype(str).str.zfill(2)
         }
     )
     route_ids = assigned_jobs["Vehicle Id"].unique()
